+++ analysis/import/SCMLogParser/src/main/java/de/maibornwolff/codecharta/importer/scmlogparser/parser/CommitCollector.java
@@ -2,6 +2,7 @@ package de.maibornwolff.codecharta.importer.scmlogparser.parser;
 
 import de.maibornwolff.codecharta.model.input.Commit;
 import de.maibornwolff.codecharta.model.input.VersionControlledFile;
+import de.maibornwolff.codecharta.model.input.metrics.MetricsFactory;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,32 +11,29 @@ import java.util.stream.Collector;
 
 class CommitCollector {
 
-    static Collector<Commit, ?, List<VersionControlledFile>> create() {
-        CommitCollector collector = new CommitCollector();
+    MetricsFactory metricsFactory;
+
+    private CommitCollector(MetricsFactory metricsFactory) {
+        this.metricsFactory = metricsFactory;
+    }
+
+    static Collector<Commit, ?, List<VersionControlledFile>> create(MetricsFactory metricsFactory) {
+        CommitCollector collector = new CommitCollector(metricsFactory);
         return Collector.of(ArrayList::new, collector::collectCommit, collector::combineForParallelExecution);
     }
 
     private void collectCommit(List<VersionControlledFile> versionControlledFiles, Commit commit) {
-        removeEmptyFiles(commit);
-        if (isEmpty(commit)) {
+        if (commit.isEmpty()) {
             return;
         }
         addYetUnknownFilesToVersionControlledFiles(versionControlledFiles, commit.getFilenames());
         addCommitMetadataToMatchingVersionControlledFiles(commit, versionControlledFiles);
     }
 
-    private void removeEmptyFiles(Commit commit) {
-        commit.getFilenames().removeIf(String::isEmpty);
-    }
-
-    private boolean isEmpty(Commit commit) {
-        return commit.getFilenames().isEmpty();
-    }
-
     private void addYetUnknownFilesToVersionControlledFiles(List<VersionControlledFile> versionControlledFiles, List<String> filenamesOfCommit) {
         filenamesOfCommit.stream()
-            .filter(filename -> !versionControlledFilesContainsFile(versionControlledFiles, filename))
-            .forEach(unknownFilename-> addYetUnknownFile(versionControlledFiles, unknownFilename));
+                .filter(filename -> !versionControlledFilesContainsFile(versionControlledFiles, filename))
+                .forEach(unknownFilename -> addYetUnknownFile(versionControlledFiles, unknownFilename));
     }
 
     private boolean versionControlledFilesContainsFile(List<VersionControlledFile> versionControlledFiles, String filename) {
@@ -44,20 +42,20 @@ class CommitCollector {
 
     private Optional<VersionControlledFile> findVersionControlledFileByFilename(List<VersionControlledFile> versionControlledFiles, String filename) {
         return versionControlledFiles.stream()
-            .filter(commit -> commit.getFilename().equals(filename))
-            .findFirst();
+                .filter(commit -> commit.getFilename().equals(filename))
+                .findFirst();
     }
 
     private boolean addYetUnknownFile(List<VersionControlledFile> versionControlledFiles, String filenameOfYetUnversionedFile) {
-        VersionControlledFile missingVersionControlledFile = new VersionControlledFile(filenameOfYetUnversionedFile);
+        VersionControlledFile missingVersionControlledFile = new VersionControlledFile(filenameOfYetUnversionedFile, metricsFactory);
         return versionControlledFiles.add(missingVersionControlledFile);
     }
 
     private void addCommitMetadataToMatchingVersionControlledFiles(Commit commit, List<VersionControlledFile> versionControlledFiles) {
         commit.getFilenames().stream()
                 .map(file -> findVersionControlledFileByFilename(versionControlledFiles, file))
-                .filter(Optional::isPresent)
-                .forEach(vcFile -> vcFile.get().registerCommit(commit));
+                .flatMap(Optional::stream)
+                .forEach(vcFile -> vcFile.registerCommit(commit));
 
     }